[receiver/k8sobjects] Fix issue where resourceVersion was not being remembered, (#24806)
**Description:** <Describe what has changed.>
Fixes an issue where the latest resourceVersion that `RetryWatcher` is
remembering is not honored.

It turns out that `RetryWatcher` also doesn't like being passed `""` or
`"0"` during startup, and if you pass it a default value of `"1"` it
will most likely return with a `410` complaining that the resource is
too old and is long gone.

To fix this, we not perform an initial `List` to get the latest values
and a proper resourceVersion. This is recommended by
https://kubernetes.io/docs/reference/using-api/api-concepts/#efficient-detection-of-changes.

**Link to tracking Issue:**
Closes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/24805

**Testing:**
I tested this a lot manually. I updated the unit tests, but it was a
struggle - the fake client framework doesn't honor the resourceVersion
concept like k8s actually does.

**Documentation:** 
Updated the readme to call out that Watch mode needs list access if no
resource_version is provided.